requirements.txt
termcolor==1.1.0
Flask==1.1.1
python-dotenv==0.10.3
Flask-Pymongo==2.3.0